Using Read and Write open parameter instead of Write only
When the user switch to the socket connection, MemTrace needs to read back the content of the file, the file needs to be opened with the O_RDWR  in order to ensure the right operation order:

O_RDONLY
Open for reading only.
O_WRONLY
Open for writing only.
O_RDWR
Open for reading and writing. The result is undefined if this flag is applied to a FIFO.